requirements.txt
dify_plugin>=0.2.0,<0.3.0
google-cloud-aiplatform==1.106.0
